Reworked ticket transfere. Some small fixes still needed.

+++ tickle/views/products.py
@@ -1,18 +1,19 @@
 # -*- coding: utf-8 -*-
 from django.utils.translation import ugettext_lazy as _
-from django.views.generic import ListView, DeleteView, UpdateView
+from django.views.generic import ListView, DeleteView, UpdateView, FormView
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect
 from django.core.urlresolvers import reverse_lazy
 from django.core.exceptions import PermissionDenied
 from django.contrib import messages
 from django.http import Http404
+from django.db.transaction import atomic
 
 from datetime import datetime
 
 from guardian.mixins import LoginRequiredMixin
 
-from tickle.forms import IdentifyForm
+from tickle.forms import IdentifyForm, SearchPersonForm
 from tickle.models.products import Holding, Purchase, Product, ShoppingCart
 from tickle.utils.mail import TemplatedEmail
 
@@ -111,12 +112,12 @@ def complete_purchase(request):
 
 class ExchangeView(LoginRequiredMixin, UpdateView):
     model = Holding
-    form_class = IdentifyForm
-    template_name = 'tickle/exchange.html'
+    form_class = SearchPersonForm
+    template_name = 'tickle/transfer.html'
 
     def verify(self, holding):
         person = self.request.user.person
-        if holding.person != person:  # or holding.purchase.person != person:
+        if holding.person != person or holding.purchase.person != person or not holding.transferable:
             raise PermissionDenied
 
     def get(self, request, *args, **kwargs):
@@ -127,11 +128,68 @@ class ExchangeView(LoginRequiredMixin, UpdateView):
         holding = self.get_object()
         self.verify(holding)
 
-        holding.person = form.get_existing_person_or_none()
-
-        # Recalculate discounts.
-        holding.holding_discounts.all().delete()
-        holding.product.product_discounts.eligible(holding.person).copy_to_holding_discounts(holding)
-
+        person = form.get_person()
+        holding.transferee = person
         holding.save()
-        return redirect('tickle:exchange_ticket_success')
+
+        msg = TemplatedEmail(
+            to=[person.pretty_email],
+            from_email='Biljett SOF15 <biljett@sof15.se>',
+            subject_template='tickle/email/transfer_subject.txt',
+            body_template_html='tickle/email/transfer.html',
+            context={
+                'holding': holding,
+            },
+            tags=['tickle', 'ticket', 'transfer'])
+        msg.send()
+        return redirect('tickle:transfer_ticket_success')
+
+
+class ConfirmExchangeView(LoginRequiredMixin, UpdateView):
+    model = Holding
+    fields = []
+    template_name = 'tickle/transfer_confirm.html'
+
+    def verify(self, holding):
+        person = self.request.user.person
+        if holding.transferee != person or not holding.transferable:
+            raise PermissionDenied
+
+    def get(self, request, *args, **kwargs):
+        self.verify(self.get_object())
+        return super(ConfirmExchangeView, self).get(request, *args, **kwargs)
+
+    def form_valid(self, form):
+        holding = self.get_object()
+        self.verify(holding)
+
+        if '_confirm' in self.request.POST:
+            transferee = holding.transferee
+            person = holding.person
+            purchase = holding.purchase
+            discounted_total = holding.discounted_total
+
+            with atomic():
+                holding.person = transferee
+                holding.transferee = None
+                if purchase.holdings.count() > 1:
+                    holding.purchase = Purchase.objects.create(person=transferee, purchased=purchase.purchased)  # Now or previous purchase time?
+                else:
+                    purchase.person = transferee
+                    purchase.save()
+                    holding.purchase = purchase
+                # Recalculate discounts.
+                holding.holding_discounts.all().delete()
+                holding.product.product_discounts.eligible(transferee).copy_to_holding_discounts(holding)
+                holding.save()
+
+                if discounted_total - holding.discounted_total != 0:
+                    # Price difference, invoice original owner with the difference.
+                    # TODO: Make it properly working.
+                    Purchase.objects.create(person=person, purchase=datetime.now())
+
+            return redirect('tickle:transfer_ticket_confirm_success')
+        elif '_decline' in self.request.POST:
+            holding.transferee = None
+            holding.save()
+        return redirect('profile', pk=self.request.user.person.pk)
